requirements.txt
streamlit
pdfplumber
pytesseract
pdf2image
pillow
openai
pandas
docx
python-docx
shutil
json
os
